Add chart display mode
