+++ lib/l10n/app_de.arb
@@ -84,6 +84,8 @@
   "cantUpdateLogsPerQuery": "Anzahl der Logs pro Anfrage konnte nicht aktualisiert werden",
   "changeServer": "Server ändern",
   "changeTheme": "App-Design ändern",
+  "chartDisplayModeSubtitle": "Wählen Sie das Diagrammformat für den Start- und Statistikbildschirm aus",
+  "chartDisplayModeTitle": "Diagramm-Anzeigemodus",
   "charts": "Diagramme",
   "chartsNotLoaded": "Diagramme konnten nicht geladen werden.",
   "checkAddress": "Fehlgeschlagen. Überprüfen Sie die Adresse.",
@@ -154,8 +156,6 @@
   "domainUpdated": "Domain wurde erfolgreich aktualisiert",
   "domains": "Domains",
   "domainsAdlists": "Domains in Adlist",
-  "domainsClientsDataMode": "Statistik-Anzeigemodus",
-  "domainsClientsDataModeDescription": "Ändert die Art, wie Domänen- und Client-Daten angezeigt werden",
   "domainsNotLoaded": "Domain-Liste konnte nicht geladen werden",
   "dontCheckCertificate": "SSL Zertifikat nicht überprüfen",
   "dontCheckCertificateDescription": "Überschreibt die SSL-Zertifikatsvalidierung des Servers",
@@ -204,6 +204,8 @@
   "hideZeroValuesDescription": "Versteckt Nullwerte aus dem Client-Diagramm",
   "hits": "Treffer:",
   "home": "Home",
+  "homeChartModeDescription": "Wählen Sie aus, wie die 24-Stunden-Aktivität auf dem Startbildschirm dargestellt wird.",
+  "homeChartModeTitle": "Startbildschirm",
   "host": "Host",
   "hour1": "1 Stunde",
   "hours": "Stunden",
@@ -377,6 +379,8 @@
   "sslErrorShort": "SSL Fehler.",
   "startTimeNotBeforeEndTime": "Die Startzeit liegt nicht vor der Endzeit",
   "statistics": "Statistiken",
+  "statsChartModeDescription": "Wählen Sie aus, wie Statistiken für Domains und Clients angezeigt werden.",
+  "statsChartModeTitle": "Statistikbildschirm",
   "statsNotLoaded": "Statistiken konnten nicht geladen werden.",
   "status": "Status",
   "statusDisabled": "Status: Inaktiv",
@@ -407,8 +411,6 @@
   "totalQueries24": "Gesamte Abfragen der letzten 24h",
   "tryAgainLater": "Bitte versuchen Sie es später noch einmal.",
   "tryReconnect": "Erneut verbinden",
-  "twentyFourHourViewMode": "24-Stunden-Ansichtsmodus",
-  "twentyFourHourViewModeSubtitle": "Ändert die Anzeigeoptionen für den Home-Bildschirm",
   "type": "Typ",
   "uncheckAll": "Alle deaktivieren",
   "unknown": "Unbekannt",
