+++ lib/l10n/app_en.arb
@@ -84,6 +84,8 @@
   "cantUpdateLogsPerQuery": "Logs quantity per request couldn't be updated",
   "changeServer": "Change server",
   "changeTheme": "Change app theme",
+  "chartDisplayModeSubtitle": "Choose how graphs are displayed on the Home and Statistics screens",
+  "chartDisplayModeTitle": "Chart Display Mode",
   "charts": "Charts",
   "chartsNotLoaded": "Charts could not be loaded",
   "checkAddress": "Failed. Check address.",
@@ -154,8 +156,6 @@
   "domainUpdated": "Domain updated successfully",
   "domains": "Domains",
   "domainsAdlists": "Domains on Adlists",
-  "domainsClientsDataMode": "Statistics display mode",
-  "domainsClientsDataModeDescription": "Changes the way that domains and clients data is displayed",
   "domainsNotLoaded": "Domains list couldn't be loaded",
   "dontCheckCertificate": "Don't check SSL certificate",
   "dontCheckCertificateDescription": "Overrides the server's SSL certificate validation",
@@ -204,6 +204,8 @@
   "hideZeroValuesDescription": "Hides zero values from clients chart",
   "hits": "Hits:",
   "home": "Home",
+  "homeChartModeDescription": "Choose how the 24-hour activity is visualized on the Home screen.",
+  "homeChartModeTitle": "Home Screen",
   "host": "Host",
   "hour1": "1 hour",
   "hours": "hours",
@@ -377,6 +379,8 @@
   "sslErrorShort": "SSL error.",
   "startTimeNotBeforeEndTime": "Start time is not before end time",
   "statistics": "Statistics",
+  "statsChartModeDescription": "Choose how statistics are displayed for domains and clients.",
+  "statsChartModeTitle": "Statistics Screen",
   "statsNotLoaded": "Stats could not be loaded",
   "status": "Status",
   "statusDisabled": "Status: disabled",
@@ -407,8 +411,6 @@
   "totalQueries24": "Total queries last 24 hours",
   "tryAgainLater": "Please try again later.",
   "tryReconnect": "Try reconnect",
-  "twentyFourHourViewMode": "24-Hour View Mode",
-  "twentyFourHourViewModeSubtitle": "Changes the display mode for the Home screen",
   "type": "Type",
   "uncheckAll": "Uncheck all",
   "unknown": "Unknown",
