+++ lib/l10n/app_es.arb
@@ -84,6 +84,8 @@
   "cantUpdateLogsPerQuery": "No se ha podido actualizar la cantidad de registros por petición.",
   "changeServer": "Cambiar servidor",
   "changeTheme": "Cambiar tema de la app",
+  "chartDisplayModeSubtitle": "Elige cómo se muestran los gráficos en la pantalla de inicio y estadísticas",
+  "chartDisplayModeTitle": "Modo de visualización de gráficos",
   "charts": "Gráficos",
   "chartsNotLoaded": "No se han podido cargar los gráficos",
   "checkAddress": "Error. Comprueba la dirección.",
@@ -155,8 +157,6 @@
   "domainUpdated": "Dominio actualizado correctamente",
   "domains": "Dominios",
   "domainsAdlists": "Dominios en Adlists",
-  "domainsClientsDataMode": "Visualización de estadísticas",
-  "domainsClientsDataModeDescription": "Cambia la manera en la que se muestran los datos de clientes y dominios",
   "domainsNotLoaded": "No se ha podido cargar la lista de dominios",
   "dontCheckCertificate": "No comprobar el certificado SSL",
   "dontCheckCertificateDescription": "Anula la validación del certificado SSL del servidor",
@@ -205,6 +205,8 @@
   "hideZeroValuesDescription": "Oculta los valores a cero del gráfico de clientes",
   "hits": "Veces:",
   "home": "Inicio",
+  "homeChartModeDescription": "Elige cómo se visualiza la actividad de 24 horas en la pantalla de inicio.",
+  "homeChartModeTitle": "Pantalla de inicio",
   "host": "Anfitrión",
   "hour1": "1 hora",
   "hours": "horas",
@@ -378,6 +380,8 @@
   "sslErrorShort": "Error de SSL.",
   "startTimeNotBeforeEndTime": "Tiempo de inicio no es anterior a tiempo de fin",
   "statistics": "Estadísticas",
+  "statsChartModeDescription": "Elige cómo se muestran las estadísticas de dominios y clientes.",
+  "statsChartModeTitle": "Pantalla de estadísticas",
   "statsNotLoaded": "No se han podido cargar las estadísticas",
   "status": "Estado",
   "statusDisabled": "Estado: desactivado",
@@ -408,8 +412,6 @@
   "totalQueries24": "Total de peticiones en últimas 24 horas",
   "tryAgainLater": "Por favor, inténtelo de nuevo más tarde.",
   "tryReconnect": "Intentar reconectar",
-  "twentyFourHourViewMode": "Modo de vista de 24 horas",
-  "twentyFourHourViewModeSubtitle": "Cambia el modo de visualización de la pantalla de inicio",
   "type": "Tipo",
   "uncheckAll": "Deseleccionar todos",
   "unknown": "Desconocido",
