+++ lib/l10n/app_ja.arb
@@ -84,6 +84,8 @@
   "cantUpdateLogsPerQuery": "ログ数のリクエストを更新できません",
   "changeServer": "サーバーを変更",
   "changeTheme": "アプリテーマを変更",
+  "chartDisplayModeSubtitle": "ホーム画面と統計画面で表示されるグラフの形式を選択できます",
+  "chartDisplayModeTitle": "グラフ表示モード",
   "charts": "チャート",
   "chartsNotLoaded": "チャートを読み込めませんでした",
   "checkAddress": "失敗しました。アドレスを確認してください。",
@@ -154,8 +156,6 @@
   "domainUpdated": "ドメインの更新に成功しました",
   "domains": "ドメイン",
   "domainsAdlists": "Adlists上のドメイン",
-  "domainsClientsDataMode": "統計表示モード",
-  "domainsClientsDataModeDescription": "ドメインとクライアントデータの表示方法を変更します",
   "domainsNotLoaded": "ドメインリストを読み込めませんでした",
   "dontCheckCertificate": "SSL証明書を確認しない",
   "dontCheckCertificateDescription": "サーバーのSSL証明書の検証を無効にします",
@@ -204,6 +204,8 @@
   "hideZeroValuesDescription": "クライアントチャートからゼロ値を非表示にします",
   "hits": "ヒット：",
   "home": "ホーム",
+  "homeChartModeDescription": "ホーム画面に表示する24時間のアクティビティの形式を選択してください。",
+  "homeChartModeTitle": "ホーム画面",
   "host": "ホスト",
   "hour1": "1時間",
   "hours": "時間",
@@ -377,6 +379,8 @@
   "sslErrorShort": "SSLエラー。",
   "startTimeNotBeforeEndTime": "開始時間が終了時間よりも前である必要があります",
   "statistics": "統計",
+  "statsChartModeDescription": "ドメインとクライアントの統計情報の表示形式を選択してください。",
+  "statsChartModeTitle": "統計画面",
   "statsNotLoaded": "統計を読み込めませんでした",
   "status": "ステータス",
   "statusDisabled": "ステータス：無効",
@@ -407,8 +411,6 @@
   "totalQueries24": "過去24時間の総クエリ数",
   "tryAgainLater": "しばらくしてから再度お試しください",
   "tryReconnect": "再接続を試みる",
-  "twentyFourHourViewMode": "24時間表示モード",
-  "twentyFourHourViewModeSubtitle": "Home画面の表示方法を変更します",
   "type": "タイプ",
   "uncheckAll": "すべて選択解除",
   "unknown": "不明",
