+++ lib/l10n/app_pl.arb
@@ -84,6 +84,8 @@
   "cantUpdateLogsPerQuery": "Nie można zaktualizować liczby dzienników na żądanie.",
   "changeServer": "Zmień serwer",
   "changeTheme": "Zmień motyw aplikacji",
+  "chartDisplayModeSubtitle": "Wybierz sposób wyświetlania wykresów na ekranie głównym i statystyk",
+  "chartDisplayModeTitle": "Tryb wyświetlania wykresu",
   "charts": "Wykresy",
   "chartsNotLoaded": "Nie udało się wczytać wykresów",
   "checkAddress": "Niepowodzenie. Sprawdź adres IP.",
@@ -154,8 +156,6 @@
   "domainUpdated": "Domen zaktualizowano pomyślnie",
   "domains": "Domeny",
   "domainsAdlists": "Domeny na listach reklamowych",
-  "domainsClientsDataMode": "Tryb wyświetlania statystyk",
-  "domainsClientsDataModeDescription": "Zmienia sposób wyświetlania danych domen i klientów",
   "domainsNotLoaded": "Nie można załadować listy domen",
   "dontCheckCertificate": "Nie sprawdzaj certyfikatu SSL",
   "dontCheckCertificateDescription": "Zastępuje sprawdzanie poprawności certyfikatu SSL serwera",
@@ -204,6 +204,8 @@
   "hideZeroValuesDescription": "Ukrywa wartości zerowe na wykresie klientów",
   "hits": "Trafienia:",
   "home": "Główna",
+  "homeChartModeDescription": "Wybierz sposób wyświetlania aktywności z ostatnich 24 godzin na ekranie głównym.",
+  "homeChartModeTitle": "Ekran główny",
   "host": "Host",
   "hour1": "1 godzina",
   "hours": "Godziny",
@@ -377,6 +379,8 @@
   "sslErrorShort": "Błąd SSL.",
   "startTimeNotBeforeEndTime": "Czas rozpoczęcia nie jest wcześniejszy niż czas zakończenia",
   "statistics": "Statystyki",
+  "statsChartModeDescription": "Wybierz sposób wyświetlania statystyk dla domen i klientów.",
+  "statsChartModeTitle": "Ekran statystyk",
   "statsNotLoaded": "Nie udało się załadować statystyk",
   "status": "Status",
   "statusDisabled": "Stan: Wyłączony",
@@ -407,8 +411,6 @@
   "totalQueries24": "Łączna liczba zapytań w ciągu ostatnich 24 godzin",
   "tryAgainLater": "Spróbuj ponownie później.",
   "tryReconnect": "Spróbuj połączyć się ponownie",
-  "twentyFourHourViewMode": "Tryb widoku 24-godzinnego",
-  "twentyFourHourViewModeSubtitle": "Zmienia sposób wyświetlania ekranu głównego",
   "type": "Typ",
   "uncheckAll": "Odznacz wszystko",
   "unknown": "Nieznany",
