+++ lib/screens/settings/app_settings/advanced_settings/advanced_options.dart
@@ -15,11 +15,10 @@ import 'package:pi_hole_client/providers/servers_provider.dart';
 import 'package:pi_hole_client/screens/app_logs/app_logs.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/app_unlock_setup_modal.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/auto_refresh_time_screen.dart';
+import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/chart_visualization_screen.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/enter_passcode_modal.dart';
-import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/home_visualization_screen.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/logs_quantity_load_screen.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/reset_screen.dart';
-import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/statistics_visualization_screen.dart';
 import 'package:pi_hole_client/widgets/custom_list_tile.dart';
 import 'package:pi_hole_client/widgets/section_label.dart';
 import 'package:provider/provider.dart';
@@ -257,31 +256,13 @@ class AdvancedOptions extends StatelessWidget {
             ),
             CustomListTile(
               leadingIcon: Icons.pie_chart_rounded,
-              label: AppLocalizations.of(context)!.domainsClientsDataMode,
-              description: AppLocalizations.of(context)!
-                  .domainsClientsDataModeDescription,
-              onTap: () => Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => const StatisticsVisualizationScreen(),
-                ),
-              ),
-              padding: const EdgeInsets.only(
-                top: 10,
-                bottom: 10,
-                left: 20,
-                right: 10,
-              ),
-            ),
-            CustomListTile(
-              leadingIcon: Icons.bar_chart_rounded,
-              label: AppLocalizations.of(context)!.twentyFourHourViewMode,
+              label: AppLocalizations.of(context)!.chartDisplayModeTitle,
               description:
-                  AppLocalizations.of(context)!.twentyFourHourViewModeSubtitle,
+                  AppLocalizations.of(context)!.chartDisplayModeSubtitle,
               onTap: () => Navigator.push(
                 context,
                 MaterialPageRoute(
-                  builder: (context) => const HomeVisualizationScreen(),
+                  builder: (context) => const ChartVisualizationScreen(),
                 ),
               ),
               padding: const EdgeInsets.only(
