+++ test/widgets/screens/settings/app_settings/advanced_settings/advanced_options_test.dart
@@ -3,8 +3,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:pi_hole_client/screens/app_logs/app_logs.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/advanced_options.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/app_unlock_setup_modal.dart';
+import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/chart_visualization_screen.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/reset_screen.dart';
-import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/statistics_visualization_screen.dart';
 
 import '../../../../helpers.dart';
 
@@ -183,7 +183,7 @@ void main() async {
           expect(find.text('Statistics display mode'), findsOneWidget);
           await tester.tap(find.text('Statistics display mode'));
           await tester.pumpAndSettle();
-          expect(find.byType(StatisticsVisualizationScreen), findsOneWidget);
+          expect(find.byType(ChartVisualizationScreen), findsOneWidget);
           expect(find.text('Statistics display mode'), findsOneWidget);
         },
       );
