+++ test/widgets/screens/settings/app_settings/advanced_settings/statistics_visualization_screen_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/advanced_options.dart';
-import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/statistics_visualization_screen.dart';
+import 'package:pi_hole_client/screens/settings/app_settings/advanced_settings/chart_visualization_screen.dart';
 
 import '../../../../helpers.dart';
 
@@ -31,11 +31,11 @@ void main() async {
 
           await tester.pumpWidget(
             testSetup.buildTestWidget(
-              const StatisticsVisualizationScreen(),
+              const ChartVisualizationScreen(),
             ),
           );
 
-          expect(find.byType(StatisticsVisualizationScreen), findsOneWidget);
+          expect(find.byType(ChartVisualizationScreen), findsOneWidget);
           expect(find.text('Statistics display mode'), findsOneWidget);
 
           expect(find.text('List'), findsOneWidget);
